Feat/better mobile

+++ mobile/app/src/main/java/com/example/test/Cash.kt
@@ -1,3 +1,3 @@
 package com.example.test
 
-class Cash(val accountNumber: String, val accountName: String, var amount: Float?)
+class Cash(val accountNumber: String, val accountName: String, var amount: Double?)
